requirements.txt
streamlit
tensorflow
numpy
Pillow
keras
python-dotenv
langchain
langchain-google-genai
google-generativeai
chromadb